Move confighandle.node package to config package.
