+++ src/main/java/org/mvplugins/multiverse/core/config/CoreConfigNodes.java
@@ -9,10 +9,10 @@ import org.bukkit.plugin.PluginManager;
 import org.jetbrains.annotations.NotNull;
 import org.mvplugins.multiverse.core.command.MVCommandManager;
 import org.mvplugins.multiverse.core.command.queue.ConfirmMode;
-import org.mvplugins.multiverse.core.confighandle.node.ConfigHeaderNode;
-import org.mvplugins.multiverse.core.confighandle.node.ConfigNode;
-import org.mvplugins.multiverse.core.confighandle.node.Node;
-import org.mvplugins.multiverse.core.confighandle.node.NodeGroup;
+import org.mvplugins.multiverse.core.config.node.ConfigHeaderNode;
+import org.mvplugins.multiverse.core.config.node.ConfigNode;
+import org.mvplugins.multiverse.core.config.node.Node;
+import org.mvplugins.multiverse.core.config.node.NodeGroup;
 import org.mvplugins.multiverse.core.destination.DestinationsProvider;
 import org.mvplugins.multiverse.core.destination.core.WorldDestination;
 import org.mvplugins.multiverse.core.event.MVDebugModeEvent;
