+++ src/main/java/org/mvplugins/multiverse/core/config/handle/CommentedConfigurationHandle.java
@@ -10,9 +10,9 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import org.mvplugins.multiverse.core.config.migration.ConfigMigrator;
-import org.mvplugins.multiverse.core.confighandle.node.CommentedNode;
-import org.mvplugins.multiverse.core.confighandle.node.NodeGroup;
-import org.mvplugins.multiverse.core.confighandle.node.ValueNode;
+import org.mvplugins.multiverse.core.config.node.CommentedNode;
+import org.mvplugins.multiverse.core.config.node.NodeGroup;
+import org.mvplugins.multiverse.core.config.node.ValueNode;
 
 /**
  * Configuration handle for commented YAML files.
