+++ src/main/java/org/mvplugins/multiverse/core/config/handle/ConfigurationSectionHandle.java
@@ -8,7 +8,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import org.mvplugins.multiverse.core.config.migration.ConfigMigrator;
-import org.mvplugins.multiverse.core.confighandle.node.NodeGroup;
+import org.mvplugins.multiverse.core.config.node.NodeGroup;
 
 /**
  * Configuration handle for a single configuration section.
