+++ src/main/java/org/mvplugins/multiverse/core/config/handle/FileConfigurationHandle.java
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import org.mvplugins.multiverse.core.config.migration.ConfigMigrator;
-import org.mvplugins.multiverse.core.confighandle.node.NodeGroup;
+import org.mvplugins.multiverse.core.config.node.NodeGroup;
 
 /**
  * Generic configuration handle for file based configurations.
