+++ src/main/java/org/mvplugins/multiverse/core/config/handle/MemoryConfigurationHandle.java
@@ -6,7 +6,7 @@ import org.bukkit.configuration.MemoryConfiguration;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.mvplugins.multiverse.core.config.migration.ConfigMigrator;
-import org.mvplugins.multiverse.core.confighandle.node.NodeGroup;
+import org.mvplugins.multiverse.core.config.node.NodeGroup;
 
 import java.util.logging.Logger;
