+++ src/main/java/org/mvplugins/multiverse/core/config/node/Node.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.confighandle.node;
+package org.mvplugins.multiverse.core.config.node;
 
 import org.jetbrains.annotations.NotNull;
