+++ src/main/java/org/mvplugins/multiverse/core/world/WorldConfigNodes.java
@@ -17,9 +17,9 @@ import org.jetbrains.annotations.NotNull;
 import org.mvplugins.multiverse.core.MultiverseCore;
 import org.mvplugins.multiverse.core.config.functions.NodeSerializer;
 import org.mvplugins.multiverse.core.event.MVWorldPropertyChangeEvent;
-import org.mvplugins.multiverse.core.confighandle.node.ConfigNode;
-import org.mvplugins.multiverse.core.confighandle.node.ListConfigNode;
-import org.mvplugins.multiverse.core.confighandle.node.NodeGroup;
+import org.mvplugins.multiverse.core.config.node.ConfigNode;
+import org.mvplugins.multiverse.core.config.node.ListConfigNode;
+import org.mvplugins.multiverse.core.config.node.NodeGroup;
 import org.mvplugins.multiverse.core.economy.MVEconomist;
 import org.mvplugins.multiverse.core.utils.MaterialConverter;
 import org.mvplugins.multiverse.core.world.helpers.EnforcementHandler;
